\name{sim.info}
\alias{sim.info}
\docType{data}
\title{ Associated Tag SNPs for Simulation  }
\description{
  Matrix giving the information on the 60 SNPs in the
  \code{\link{dna.snp.full}} data set, giving the SNP rs number, the gene
  of the SNP, the assumed odds ratio of the SNP, the minor allele
  frequency of the SNP and the assumed genetic model of the SNP.
}

\usage{data(sim.info)}

\examples{
  data(sim.info)
}

\keyword{datasets}

